+++ packages/core/src/core/handleEvents.ts
@@ -1,9 +1,6 @@
 import ErrorStackParser from 'error-stack-parser';
-import { record } from 'rrweb';
 import {
   openWhiteScreen,
-  getResource,
-  getWebVitals,
   transportData,
   breadcrumb,
   resourceTransform,
@@ -17,11 +14,6 @@ import {
   getTimestamp,
   parseUrlToObj,
   unknownToString,
-  on,
-  _global,
-  _support,
-  zip,
-  generateUUID,
 } from '@websee/utils';
 import { ErrorTarget, RouteHistory, HttpData } from '@websee/types';
 
@@ -160,103 +152,6 @@ const HandleEvents = {
       }
     } catch (er) {
       // console.error('web-see: handleUnhandleRejection错误解析异常:', er);
-    }
-  },
-  handlePerformance(): void {
-    try {
-      // 获取FCP、LCP、TTFB、FID等指标
-      getWebVitals((res: any) => {
-        // name指标名称、rating 评级、value数值
-        const { name, rating, value } = res;
-        transportData.send({
-          type: EVENTTYPES.PERFORMANCE,
-          status: STATUS_CODE.OK,
-          time: getTimestamp(),
-          name,
-          rating,
-          value,
-        });
-      });
-    } catch (err) {
-      // console.err('性能指标获取错误:' + err);
-    }
-
-    const observer = new PerformanceObserver(list => {
-      for (const long of list.getEntries()) {
-        // 上报长任务详情
-        transportData.send({
-          type: EVENTTYPES.PERFORMANCE,
-          name: 'longTask',
-          longTask: long,
-          time: getTimestamp(),
-          status: STATUS_CODE.OK,
-        });
-      }
-    });
-    observer.observe({ entryTypes: ['longtask'] });
-
-    on(_global, 'load', function () {
-      // 上报资源列表
-      transportData.send({
-        type: EVENTTYPES.PERFORMANCE,
-        name: 'resource_list',
-        time: getTimestamp(),
-        status: STATUS_CODE.OK,
-        resourceList: getResource(),
-      });
-
-      // 上报内存情况, safari、firefox不支持该属性
-      if (performance.memory) {
-        transportData.send({
-          type: EVENTTYPES.PERFORMANCE,
-          name: 'memory',
-          time: getTimestamp(),
-          status: STATUS_CODE.OK,
-          memory: {
-            jsHeapSizeLimit: performance.memory.jsHeapSizeLimit,
-            totalJSHeapSize: performance.memory.totalJSHeapSize,
-            usedJSHeapSize: performance.memory.usedJSHeapSize,
-          },
-        });
-      }
-    });
-  },
-  handleScreen(): void {
-    try {
-      // events存储录屏信息
-      let events: any[] = [];
-      // 调用stopFn停止录像
-      // let stopFn = record({
-      record({
-        emit(event, isCheckout) {
-          if (isCheckout) {
-            // 此段时间内发生错误，上报录屏信息
-            if (_support.hasError) {
-              const recordScreenId = _support.recordScreenId;
-              _support.recordScreenId = generateUUID();
-              transportData.send({
-                type: EVENTTYPES.RECORDSCREEN,
-                recordScreenId,
-                time: getTimestamp(),
-                status: STATUS_CODE.OK,
-                events: zip(events),
-              });
-              events = [];
-              _support.hasError = false;
-            } else {
-              // 不上报，清空录屏
-              events = [];
-              _support.recordScreenId = generateUUID();
-            }
-          }
-          events.push(event);
-        },
-        recordCanvas: true,
-        // 默认每10s重新制作快照
-        checkoutEveryNms: 1000 * options.recordScreentime,
-      });
-    } catch (err) {
-      // console.err('录屏报错:' + err);
     }
   },
   handleWhiteScreen(): void {