+++ packages/core/src/core/index.ts
@@ -7,5 +7,4 @@ export * from './subscribe';
 export * from './handleEvents';
 export * from './replace';
 export * from './setupReplace';
-export * from './performance';
 export * from './whiteScreen';