+++ packages/core/src/core/options.ts
@@ -1,5 +1,4 @@
-import { validateOption, _support, setSilentFlag, generateUUID } from '@websee/utils';
-import { EVENTTYPES } from '@websee/common';
+import { validateOption, _support, setSilentFlag } from '@websee/utils';
 import { InitOptions } from '@websee/types';
 import { breadcrumb } from './breadcrumb';
 import { transportData } from './reportData';
@@ -8,35 +7,19 @@ export class Options {
   dsn = ''; // 监控上报接口的地址
   throttleDelayTime = 0; // click事件的节流时长
   overTime = 10; // 接口超时时长
-  silentRecordScreen = false; // 是否开启录屏
-  recordScreentime = 10; // 默认录屏时长
-  recordScreenTypeList: string[]; // 上报录屏的错误列表
-  whiteBoxElements: string[]; // // 白屏检测的容器列表
+  whiteBoxElements: string[] = ['html', 'body', '#app', '#root']; // // 白屏检测的容器列表
   silentWhiteScreen = false; // 是否开启白屏检测
   skeletonProject = false; // 项目是否有骨架屏
   filterXhrUrlRegExp: any; // 过滤的接口请求正则
   handleHttpStatus: any; // 处理接口返回的 response
   repeatCodeError = false; // 是否去除重复的代码错误，重复的错误只上报一次
-  constructor() {
-    this.recordScreenTypeList = [
-      // 录屏事件集合
-      EVENTTYPES.ERROR,
-      EVENTTYPES.UNHANDLEDREJECTION,
-      EVENTTYPES.RESOURCE,
-      EVENTTYPES.FETCH,
-      EVENTTYPES.XHR,
-    ];
-    this.whiteBoxElements = ['html', 'body', '#app', '#root']; // 白屏检测的父容器列表
-  }
+  constructor() {}
   bindOptions(options: InitOptions): void {
     const {
       dsn,
       filterXhrUrlRegExp,
       throttleDelayTime = 0,
-      silentRecordScreen = false,
       overTime = 10,
-      recordScreenTypeList,
-      recordScreentime = 10,
       silentWhiteScreen = false,
       whiteBoxElements = ['html', 'body', '#app', '#root'],
       skeletonProject = false,
@@ -47,14 +30,6 @@ export class Options {
     validateOption(throttleDelayTime, 'throttleDelayTime', 'number') &&
       (this.throttleDelayTime = throttleDelayTime);
     validateOption(overTime, 'overTime', 'number') && (this.overTime = overTime);
-    validateOption(recordScreentime, 'recordScreentime', 'number') &&
-      (this.recordScreentime = recordScreentime);
-    if (validateOption(silentRecordScreen, 'silentRecordScreen', 'boolean')) {
-      this.silentRecordScreen = silentRecordScreen;
-      _support.recordScreenId = generateUUID(); // 添加初始的recordScreenId
-    }
-    validateOption(recordScreenTypeList, 'recordScreenTypeList', 'array') &&
-      (this.recordScreenTypeList = recordScreenTypeList);
     validateOption(filterXhrUrlRegExp, 'filterXhrUrlRegExp', 'regexp') &&
       (this.filterXhrUrlRegExp = filterXhrUrlRegExp);
     validateOption(silentWhiteScreen, 'silentWhiteScreen', 'boolean') &&