+++ packages/performance/global.d.ts
@@ -1,12 +1,12 @@
+declare module '*.json' {
+  const value: any;
+  export default value;
+}
+
 declare interface Performance extends Performance {
   memory?: {
     jsHeapSizeLimit: number;
     totalJSHeapSize: number;
     usedJSHeapSize: number;
   };
-}
-
-declare module '*.json' {
-  const value: any;
-  export default value;
 }