+++ packages/types/src/index.ts
@@ -1,3 +1,3 @@
 export * from './core/option';
-export * from './core/common';
+export * from './core/base';
 export * from './core/vue';