+++ packages/utils/src/core/browser.ts
@@ -1,5 +1,3 @@
-import pako from 'pako';
-import { Base64 } from 'js-base64';
 import { setFlag, _support } from './global';
 import { InitOptions } from '@websee/types';
 import { EVENTTYPES } from '@websee/common';
@@ -48,23 +46,6 @@ export function parseUrlToObj(url: string) {
   };
 }
 
-// 压缩
-export function zip(data: any): string {
-  if (!data) return data;
-  // 判断数据是否需要转为JSON
-  const dataJson =
-    typeof data !== 'string' && typeof data !== 'number' ? JSON.stringify(data) : data;
-  // 使用Base64.encode处理字符编码，兼容中文
-  const str = Base64.encode(dataJson as string);
-  const binaryString = pako.gzip(str);
-  const arr = Array.from(binaryString);
-  let s = '';
-  arr.forEach((item: number) => {
-    s += String.fromCharCode(item);
-  });
-  return Base64.btoa(s);
-}
-
 export function setSilentFlag(paramOptions: InitOptions): void {
   setFlag(EVENTTYPES.XHR, !!paramOptions.silentXhr);
   setFlag(EVENTTYPES.FETCH, !!paramOptions.silentFetch);
@@ -73,8 +54,6 @@ export function setSilentFlag(paramOptions: InitOptions): void {
   setFlag(EVENTTYPES.ERROR, !!paramOptions.silentError);
   setFlag(EVENTTYPES.HASHCHANGE, !!paramOptions.silentHashchange);
   setFlag(EVENTTYPES.UNHANDLEDREJECTION, !!paramOptions.silentUnhandledrejection);
-  setFlag(EVENTTYPES.PERFORMANCE, !!paramOptions.silentPerformance);
-  setFlag(EVENTTYPES.RECORDSCREEN, !paramOptions.silentRecordScreen);
   setFlag(EVENTTYPES.WHITESCREEN, !paramOptions.silentWhiteScreen);
 }