+++ packages/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @websee/common
 
+## 1.1.1
+
+### Patch Changes
+
+- 1.1.1
+
 ## 1.1.0
 
 ### Minor Changes