+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @websee/utils
 
+## 1.1.1
+
+### Patch Changes
+
+- 1.1.1
+- Updated dependencies
+  - @websee/common@1.1.1
+  - @websee/types@1.1.1
+  - @websee/core@1.1.1
+
 ## 1.1.0
 
 ### Minor Changes